For EXP-2012: Move tests for evaluator into their own file
