+++ components/nimbus/src/tests/mod.rs
@@ -4,6 +4,7 @@
 
 mod test_enrollment;
 mod test_enrollment_bw_compat;
+mod test_evaluator;
 mod test_lib;
 mod test_lib_bw_compat;
 mod test_lib_schema_deserialization;
